+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 ## master
 
 - Added `.introspectSplitViewController()` on iOS
+- Fixed iPad tests
+- Added iPad to CI
 - Added `.introspectColorWell()` on iOS and macOS
 - Added `.introspectButton()` on macOS
 - Fix UITextField with cornerRadius